+++ Makefile
@@ -39,6 +39,9 @@ SRC = src/shell.c \
 	src/utility/fusion.c	\
 	src/utility/split_commands.c \
 	src/utility/get_return.c	\
+	src/quotes.c \
+	src/utility/replace_var.c	\
+	src/utility/get_inhibitor.c	\
 
 OBJ = $(SRC:%.c=%.o)
 OBJ += src/main.o
