+++ include/shell.h
@@ -70,9 +70,10 @@ char *get_alias_command(char *cmd, alias_t *alias);
 char **remove_quotes(char **argv);
 bool matched_quotes(char *cmd);
 char **split_args(char *cmd, char **argv);
-char * get_var(char *cmd, env_t *env);
+char * replace_var(char *cmd, env_t *env);
 char *get_inhibitor(char *cmd);
 char *get_var_value(char *var, env_t *env);
+char *isolate_var(char *cmd, env_t *env);
 
 
 #define ERROR 84
